+++ packages/@markuplint/svelte-parser/src/parser.ts
@@ -1,5 +1,10 @@
 import type { SvelteAwaitBlock, SvelteEachBlock, SvelteIfBlock, SvelteNode } from './svelte-parser/index.js';
-import type { MLASTParentNode, MLASTPreprocessorSpecificBlock, MLASTBlockBehavior } from '@markuplint/ml-ast';
+import type {
+	MLASTNodeTreeItem,
+	MLASTParentNode,
+	MLASTPreprocessorSpecificBlock,
+	MLASTBlockBehavior,
+} from '@markuplint/ml-ast';
 import type { ChildToken, ParseOptions, Token } from '@markuplint/parser-utils';
 
 import { ParserError, Parser, AttrState } from '@markuplint/parser-utils';
@@ -15,11 +20,10 @@ export class SvelteParser extends Parser<SvelteNode> {
 			endTagType: 'xml',
 			tagNameCaseSensitive: true,
 			ignoreTags: [
-				{
-					type: 'Script',
-					start: '<script',
-					end: '</script>',
-				},
+				// Fixed(Resolved) #2505
+				// The script tag needs to be passed with the lang attribute to the parser,
+				// so instead of excluding it, we convert it to PSBlock in visitText.
+				// Keep the style tag as is for now.
 				{
 					type: 'Style',
 					start: '<style',
@@ -52,6 +56,24 @@ export class SvelteParser extends Parser<SvelteNode> {
 		}
 
 		return super.parseError(error);
+	}
+
+	visitText(token: ChildToken): readonly MLASTNodeTreeItem[] {
+		const nodes = super.visitText(token, {
+			researchTags: false,
+			invalidTagAsText: false,
+		});
+
+		return nodes.flatMap(node => {
+			if (node.nodeName === '#text' && /^<script[\s>]/i.test(node.raw)) {
+				return this.visitPsBlock({
+					...node,
+					nodeName: 'Script',
+					isFragment: false,
+				});
+			}
+			return node;
+		});
 	}
 
 	nodeize(