+++ website/src/utils/getOgImgUrl.ts
@@ -12,7 +12,8 @@ const fonts: Record<string, [label: string, title: string]> = {
 const getTextWidth = (text: string): number => {
   let count = 0;
   for (const element of text) {
-    /[\u0020-\u007E]/.test(element) ? (count += 1) : (count += 2);
+    const charLength = /[\u0020-\u007E]/.test(element) ? 1 : 2;
+    count += charLength;
   }
   return count;
 };