+++ packages/@markuplint-dev/eslint-config/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/markuplint/markuplint/compare/@markuplint-dev/eslint-config@1.0.13...@markuplint-dev/eslint-config@1.0.14) (2025-03-09)
+
+**Note:** Version bump only for package @markuplint-dev/eslint-config
+
 ## [1.0.13](https://github.com/markuplint/markuplint/compare/@markuplint-dev/eslint-config@1.0.12...@markuplint-dev/eslint-config@1.0.13) (2025-02-27)
 
 **Note:** Version bump only for package @markuplint-dev/eslint-config