+++ packages/@markuplint/pretenders/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.20](https://github.com/markuplint/markuplint/compare/@markuplint/pretenders@0.0.19...@markuplint/pretenders@0.0.20) (2025-03-09)
+
+**Note:** Version bump only for package @markuplint/pretenders
+
 ## [0.0.19](https://github.com/markuplint/markuplint/compare/@markuplint/pretenders@0.0.18...@markuplint/pretenders@0.0.19) (2025-02-27)
 
 **Note:** Version bump only for package @markuplint/pretenders