+++ packages/@markuplint/rules/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.11](https://github.com/markuplint/markuplint/compare/@markuplint/rules@4.10.10...@markuplint/rules@4.10.11) (2025-03-09)
+
+**Note:** Version bump only for package @markuplint/rules
+
 ## [4.10.10](https://github.com/markuplint/markuplint/compare/@markuplint/rules@4.10.9...@markuplint/rules@4.10.10) (2025-02-27)
 
 ### Bug Fixes